editor shortcuts, remove cacheOnly
