+++ packages/databyss-services/pages/actions.js
@@ -134,19 +134,13 @@ export function savePatch(patch) {
 export function savePage(state) {
   const id = state.page._id
   const body = cloneDeep(state)
-  if (state.updatePageInCache) {
-    // remove page title from payload to prevent name from being updated
-    delete body.page
-  }
-  delete body.updatePageInCache
+
   return dispatch => {
     dispatch({
       type: CACHE_PAGE,
       payload: { body, id },
     })
-    if (!state.updatePageInCache) {
-      services.savePage(body)
-    }
+    services.savePage(body)
   }
 }
